% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{npy_dtype}
\alias{npy_dtype}
\title{The NumPy dtype for an R value}
\usage{
npy_dtype(x)
}
\arguments{
\item{x}{An R value.}
}
\value{
A string such as \code{"<f8"}.
}
\description{
The dtype string \code{\link[=npy_encode]{npy_encode()}} writes for \code{x} when no \code{dtype} is given.
}
\examples{
npy_dtype(1:3)
npy_dtype(c(TRUE, FALSE))
npy_dtype(matrix(0, 2, 2))
}
